requirements.txt
discord.py
aiosqlite
jishaku
discord-py-slash-command
